backend/requirements.txt
fastapi
google-genai
uvicorn
